+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Switch to gpt-5-mini.
+- Use typescript autogen for backend types in frontend.
 
 ## [v0.8.1] - 01.09.2025